Refactor TwitterClient; Make Tweet Delay configurable; Change Tweet template on Dev
