+++ Api/appsettings.Development.json
@@ -11,5 +11,9 @@
   },
   "MySQL": {
     "ConnectionString": "Server=localhost;User ID=root;Password=SweetFresh06;Database=dictionary"
+  },
+  "Twitter": {
+    "TweetTemplate": "{name}: \"{meaning}\" {link}",
+    "TweetIntervalSeconds": 5
   }
 }
