+++ Api/appsettings.json
@@ -13,6 +13,7 @@
     "ConsumerKey": "your-consumer-key",
     "ConsumerSecret": "your-consumer-secret",
     "NameUrlPrefix": "https://www.yorubaname.com/entries",
-    "TweetTemplate": "New name entry: {name}, {meaning}. More here: {link}"
+    "TweetTemplate": "New name entry: {name}, {meaning}. More here: {link}",
+    "TweetIntervalSeconds": 180
   }
 }
