+++ Infrastructure/Configuration/TwitterConfig.cs
@@ -6,8 +6,9 @@
         string AccessToken,
         string AccessTokenSecret,
         string NameUrlPrefix,
-        string TweetTemplate)
+        string TweetTemplate,
+        int TweetIntervalSeconds)
     {
-        public TwitterConfig() : this("", "", "", "", "", "") { }
+        public TwitterConfig() : this("", "", "", "", "", "", default) { }
     }
 }
