+++ Infrastructure/Services/NamePostingService.cs
@@ -1,36 +1,36 @@
 ﻿using Application.Domain;
 using Application.Events;
 using Infrastructure.Configuration;
+using Infrastructure.Twitter;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using System.Collections.Concurrent;
-using Tweetinvi;
 
 namespace Infrastructure.Services
 {
     public class NamePostingService(
         ConcurrentQueue<PostPublishedNameCommand> nameQueue,
-        ITwitterClient twitterApiClient,
+        ITwitterClientV2 twitterApiClient,
         ILogger<NamePostingService> logger,
         NameEntryService nameEntryService,
         IOptions<TwitterConfig> twitterConfig) : BackgroundService
     {
         private const string TweetComposeFailure = "Failed to build tweet for name: {name}. It was not found in the database.";
         private readonly ConcurrentQueue<PostPublishedNameCommand> _nameQueue = nameQueue;
-        private readonly TweetsV2Poster _twitterApiClient = new (twitterApiClient);
+        private readonly ITwitterClientV2 _twitterApiClient = twitterApiClient;
         private readonly ILogger<NamePostingService> _logger = logger;
         private readonly NameEntryService _nameEntryService = nameEntryService;
         private readonly TwitterConfig _twitterConfig = twitterConfig.Value;
-        private const int TweetIntervalMs = 3 * 60 * 1000; // 3 minutes
 
         protected override async Task ExecuteAsync(CancellationToken stoppingToken)
         {
+            var tweetIntervalMs = _twitterConfig.TweetIntervalSeconds * 1000;
             while (!stoppingToken.IsCancellationRequested)
             {
                 if (!_nameQueue.TryDequeue(out var indexedName))
                 {
-                    await Task.Delay(TweetIntervalMs, stoppingToken);
+                    await Task.Delay(tweetIntervalMs, stoppingToken);
                     continue;
                 }
 
@@ -56,7 +56,7 @@ namespace Infrastructure.Services
                     _nameQueue.Enqueue(indexedName);
                 }
 
-                await Task.Delay(TweetIntervalMs, stoppingToken);
+                await Task.Delay(tweetIntervalMs, stoppingToken);
             }
         }
