+++ Infrastructure/Twitter/ITwitterClientV2.cs
@@ -1,0 +1,7 @@
+﻿namespace Infrastructure.Twitter
+{
+    public interface ITwitterClientV2
+    {
+        Task<TweetV2PostResponse> PostTweet(string text);
+    }
+}
